pyproject.toml
[tool.poetry]
name = "volley-platform"
version = "0.1.0"
description = "Volley Platform backend services"
authors = ["Volley Platform Team <dev@example.com>"]
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
bcrypt = "^4.1.2"
python = "^3.11"
fastapi = "0.115.2"
uvicorn = "0.30.3"
sqlalchemy = "2.0.34"
alembic = "1.13.3"
psycopg = { version = "3.2.13", extras = ["binary"] }
python-jose = "3.3.0"
passlib = "1.7.4"
python-multipart = "0.0.9"
pydantic-settings = "2.2.1"
pandas = "2.2.3"
email-validator = "2.2.0"
psycopg2-binary = "^2.9.11"
jinja2 = "^3.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
